Refactored location
Improved assigning locations within the boundary

+++ Alarm.cs
@@ -3,8 +3,8 @@ namespace alarms
     public class Alarm
     {
         public string status { get; set; }
-        public double longtitude { get; set; }
-        public double latitude { get; set; }
+        public decimal longtitude { get; set; }
+        public decimal latitude { get; set; }
         public string image { get; set; }
     }
 }
